feat: move DialBreadcrumbPathItem type to models and update imports
